+++ src/libkiter/algorithms/throughput/state.h
@@ -71,7 +71,7 @@ public:
   bool addState(State& s);
   std::list<State>::iterator getRepeatedState();
   TIME_UNIT computeThroughput();
-  int computeIdx(State& s);
+  long computeIdx(State& s);
 
 private:
   std::list<State> visitedStates;
